Extract tools for file walking
