+++ compiler-bin/src/docs/error.rs
@@ -1,0 +1,1 @@
+
