+++ compiler-bin/src/lsp.rs
@@ -4,9 +4,8 @@ pub mod event;
 pub mod extension;
 
 use std::borrow::BorrowMut;
-use std::collections::BTreeSet;
 use std::ops::{ControlFlow, Deref};
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use std::sync::Arc;
 use std::{env, fs, mem, process};
 
@@ -23,15 +22,13 @@ use async_lsp::panic::CatchUnwindLayer;
 use async_lsp::router::Router;
 use async_lsp::server::LifecycleLayer;
 use async_lsp::{ClientSocket, ResponseError};
-use globset::{Glob, GlobSetBuilder};
 use parking_lot::RwLock;
-use path_absolutize::Absolutize;
 use tokio::task;
 use tower::ServiceBuilder;
-use walkdir::WalkDir;
 
 use crate::lsp::capabilities::negotiate_position_encoding;
 use crate::lsp::error::{AnalyzerResultExt, LspError};
+use crate::walk;
 
 #[derive(Debug)]
 pub struct LspConfig {
@@ -218,60 +215,10 @@ fn initialized_manual(state: &mut State, command: &str) -> Result<(), LspError> 
     let output = command.output()?;
     let output = str::from_utf8(&output.stdout)?;
 
-    let mut files = vec![];
-    let mut globs = GlobSetBuilder::new();
-
-    // The shallowest directory each glob may match. Walking these instead of
-    // `root` lets patterns escape the workspace through parent directories,
-    // such as `../shared/src/**/*.purs`.
-    let mut walk_roots = BTreeSet::new();
-
-    for line in output.lines() {
-        let path = root.join(line);
-        if let Ok(path) = path.absolutize()
-            && let Some(path) = path.to_str()
-            && let Ok(glob) = Glob::new(path)
-        {
-            walk_roots.insert(glob_literal_base(path));
-            globs.add(glob);
-        } else {
-            files.push(path);
-        }
-    }
-
-    let globs = globs.build()?;
-
-    tracing::info!("Found {} file literals", files.len());
-    tracing::info!("Found {} glob patterns", globs.len());
-
-    let files_from_glob: BTreeSet<PathBuf> = walk_roots
-        .into_iter()
-        .flat_map(|base| WalkDir::new(base).into_iter())
-        .filter_map(|entry| Some(entry.ok()?.into_path()))
-        .filter(|path| !globs.matches(path).is_empty())
-        .collect();
-
-    files.extend(files_from_glob);
+    let walk::Walk { files, .. } = walk::walk(root, output.lines())?;
     load_files(state, &files)?;
 
     Ok(())
-}
-
-/// Extracts the longest leading run of path components with no glob syntax.
-fn glob_literal_base(pattern: &str) -> PathBuf {
-    let mut base = PathBuf::new();
-    for component in Path::new(pattern).components() {
-        if component.as_os_str().to_string_lossy().chars().any(glob_syntax_character) {
-            break;
-        }
-        base.push(component);
-    }
-    base
-}
-
-fn glob_syntax_character(character: char) -> bool {
-    matches!(character, '*' | '?' | '[' | '{')
-        || (character == '\\' && !std::path::is_separator('\\'))
 }
 
 fn initialized_spago(state: &mut State) -> Result<(), LspError> {
@@ -595,57 +542,4 @@ pub async fn async_start(config: Arc<LspConfig>) {
 pub async fn start(config: LspConfig) {
     let config = Arc::new(config);
     async_start(Arc::clone(&config)).await
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn literal_base_stops_at_the_first_wildcard() {
-        let base = glob_literal_base("/workspace/src/**/*.purs");
-        assert_eq!(base, PathBuf::from("/workspace/src"));
-    }
-
-    #[test]
-    fn literal_base_excludes_a_wildcard_in_the_final_component() {
-        let base = glob_literal_base("/workspace/src/*.purs");
-        assert_eq!(base, PathBuf::from("/workspace/src"));
-    }
-
-    #[test]
-    fn literal_base_retains_parent_directories() {
-        let base = glob_literal_base("/workspace/../shared/src/**/*.purs");
-        assert_eq!(base, PathBuf::from("/workspace/../shared/src"));
-    }
-
-    #[test]
-    fn literal_base_of_a_pattern_without_wildcards_is_the_whole_path() {
-        let base = glob_literal_base("/workspace/src/Main.purs");
-        assert_eq!(base, PathBuf::from("/workspace/src/Main.purs"));
-    }
-
-    #[test]
-    fn literal_base_recognises_every_metacharacter() {
-        for pattern in ["/a/b/*.purs", "/a/b/?.purs", "/a/b/[abc].purs", "/a/b/{x,y}.purs"] {
-            assert_eq!(glob_literal_base(pattern), PathBuf::from("/a/b"), "pattern: {pattern}");
-        }
-    }
-
-    #[test]
-    fn literal_base_keeps_class_closer_as_a_literal() {
-        let pattern = "/workspace/src/Main].purs";
-
-        assert!(Glob::new(pattern).is_ok());
-        assert_eq!(glob_literal_base(pattern), PathBuf::from(pattern));
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn literal_base_stops_at_backslash_escape() {
-        let pattern = "/workspace/src/\\*.purs";
-
-        assert!(Glob::new(pattern).is_ok());
-        assert_eq!(glob_literal_base(pattern), PathBuf::from("/workspace/src"));
-    }
 }
